finish header, hero, and random section
I worked on organizing the header, the hero section, and the random info section in CSS. Next step is the Quote section.
Note: I used a percentage to set the size of an image, it totally messed up flexbox functionality. Stick to whole number pixels (use min/max-width to set an images dimensions)